simplified example to just be about temperature control using range
