--- examples/slider/css/range-thermostat.css
@@ -1,173 +1,0 @@
-/* CSS Document */
-
-.range-thermostat h3 {
-  color: black;
-  font-weight: bold;
-  font-size: 150%;
-}
-
-.range-thermostat-vertical {
-  touch-action: pan-x;
-}
-
-.range-thermostat-text {
-  touch-action: pan-y;
-}
-
-.range-thermostat-vertical,
-.range-thermostat-text {
-  margin-top: 1em;
-  padding: 6px;
-  width: 15em;
-}
-
-.range-thermostat-vertical label,
-.range-thermostat-text label {
-  font-weight: bold;
-  display: block;
-}
-
-.range-thermostat-vertical .range-value {
-  margin-left: 102px;
-  font-weight: bold;
-  font-size: 200%;
-  font-family: sans-serif;
-}
-
-.range-thermostat-text .labels {
-  position: relative;
-}
-
-.range-thermostat-text .labels button {
-  margin: 2px;
-  padding: 6px;
-  border: none;
-  background: transparent;
-  display: inline-block;
-  font-size: 1em;
-}
-
-/* Range Vertical Controls */
-
-.range-thermostat-vertical input[type="range"] {
-  width: 280px;
-  height: 300px;
-  background-color: transparent;
-  -webkit-appearance: none;
-  -webkit-transform: rotate(270deg);
-  -moz-transform: rotate(270deg);
-  -o-transform: rotate(270deg);
-  -ms-transform: rotate(270deg);
-  transform: rotate(270deg);
-}
-
-.range-thermostat-vertical input[type="range"]::-webkit-slider-runnable-track {
-  background: #3071a9;
-  border: 2px solid currentColor;
-  border-radius: 4px;
-  height: 10px;
-  cursor: pointer;
-}
-
-.range-thermostat-vertical input[type="range"]::-webkit-slider-thumb {
-  margin-top: -20.45px;
-  width: 13px;
-  height: 48px;
-  background: currentColor;
-  border: 2px solid currentColor;
-  border-radius: 7.5px;
-  cursor: pointer;
-  -webkit-appearance: none;
-}
-
-.range-thermostat-vertical input[type="range"]::-moz-range-track {
-  background: #3071a9;
-  border: 2px solid currentColor;
-  border-radius: 4px;
-  height: 10px;
-  cursor: pointer;
-}
-
-.range-thermostat-vertical input[type="range"]::-moz-range-thumb {
-  width: 13px;
-  height: 48px;
-  background: currentColor;
-  border: 2px solid currentColor;
-  border-radius: 7.5px;
-  cursor: pointer;
-}
-
-/* Range Text Controls */
-
-.range-thermostat-text input[type="range"] {
-  margin: 20.25px 12px;
-  background-color: transparent;
-  -webkit-appearance: none;
-}
-
-.range-thermostat-text input[type="range"]::-webkit-slider-runnable-track {
-  background: #3071a9;
-  border: 2px solid currentColor;
-  border-radius: 4px;
-  width: 100%;
-  height: 8px;
-  cursor: pointer;
-}
-
-.range-thermostat-text input[type="range"]::-webkit-slider-thumb {
-  margin-top: -20.45px;
-  width: 13px;
-  height: 48px;
-  background: currentColor;
-  border: 2px solid currentColor;
-  border-radius: 7.5px;
-  cursor: pointer;
-  -webkit-appearance: none;
-}
-
-.range-thermostat-text input[type="range"]::-moz-range-track {
-  background: #3071a9;
-  border: 2px solid currentColor;
-  border-radius: 4px;
-  width: 100%;
-  height: 8px;
-  cursor: pointer;
-}
-
-.range-thermostat-text input[type="range"]::-moz-range-thumb {
-  width: 13px;
-  height: 48px;
-  background: currentColor;
-  border: 2px solid currentColor;
-  border-radius: 7.5px;
-  cursor: pointer;
-}
-
-/* Focus and hover styling */
-
-.range-thermostat-text .labels button:hover {
-  padding: 4px;
-  border: 2px #005a9c solid;
-  border-radius: 5px;
-}
-
-.range-thermostat-vertical input[type="range"]:focus,
-.range-thermostat-text input[type="range"]:focus {
-  outline: none;
-}
-
-.range-thermostat-vertical
-  input[type="range"]:focus::-webkit-slider-runnable-track {
-  background: #3174ad;
-}
-
-.range-thermostat-text
-  input[type="range"]:focus::-webkit-slider-runnable-track {
-  background: #3174ad;
-}
-
-input[type="range"]:focus::-moz-range-thumb,
-input[type="range"]:hover::-moz-range-thumb {
-  outline-offset: 2px;
-  outline: 2px #005a9c solid;
-}
