+++ examples/slider/range-temperature.html
@@ -1,0 +1,147 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <meta charset="utf-8" />
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <title>HTML Range Examples with aria-orientation and aria-valuetext | WAI-ARIA Authoring Practices 1.2</title>
+
+  <!--  Core js and css shared by all examples; do not modify when using this template. -->
+  <link rel="stylesheet" href="https://www.w3.org/StyleSheets/TR/2016/base.css">
+  <link rel="stylesheet" href="../css/core.css">
+  <script src="../js/examples.js"></script>
+  <script src="../js/highlight.pack.js"></script>
+  <script src="../js/app.js"></script>
+
+  <link rel="stylesheet" href="css/range-temperature.css">
+  <script src="js/range-temperature.js"></script>
+</head>
+<body>
+  <nav aria-label="Related Links" class="feedback">
+    <ul>
+      <li><a href="../../#browser_and_AT_support">Browser and Assistive Technology Support</a></li>
+      <li><a href="https://github.com/w3c/aria-practices/issues/new">Report Issue</a></li>
+      <li><a href="https://github.com/w3c/aria-practices/projects/3">Related Issues</a></li>
+      <li><a href="../../#slider">Design Pattern</a></li>
+    </ul>
+  </nav>
+  <main>
+    <h1>HTML Range Examples with aria-orientation and aria-valuetext</h1>
+    <p>
+      The following example of a temperature range slider demonstrating the using <code>aria-valuetext</code> and <code>aria-orientation</code> on an <code>input[type="range"]</code> control.
+      The desired temperature is set with the range control, which is vertically oriented.
+      The range control illustrate how to use <code>aria-valuetext</code> to provide assistive technologies with meaningful names for numeric values.
+    </p>
+    <p>Similar examples include: </p>
+    <ul>
+      <li><a href="slider-color-viewer.html">Color Viewer Slider Example</a>: Basic  horizontal sliders that illustrate setting numeric values for a color picker.</li>
+      <li><a href="examples/slider/slider-temperature.html">Temperature Selector Slider Example</a>: Demonstrates using <code>aria-orientation</code> to specify vertical orientation and <code>aria-valuetext</code> to communicate unit of measure for a temperature input.</li>
+      <li><a href="slider-rating.html">Rating Slider Example</a>: Horizontal slider that demonstrates using <code>aria-valuetext</code> to communicate current and maximum value of a rating input for a five star rating scale.</li>
+      <li><a href="multithumb-slider.html">Horizontal Multi-Thumb Slider Example</a>: Demonstrates using sliders with two thumbs to provide inputs for  numeric ranges, such as for searching in a price range.</li>
+    </ul>
+    <section>
+      <div class="example-header">
+        <h2 id="ex_label">Example</h2>
+      </div>
+      <div role="separator" id="ex_start_sep" aria-labelledby="ex_start_sep ex_label" aria-label="Start of"></div>
+      <div id="ex1"
+        class="example">
+          <div class="range-temperature">
+            <label for="id-temp-range">
+              Temperature
+            </label>
+            <div class="range-value">25.0°C</div>
+            <input type="range"
+              id="id-temp-range"
+              class="veritcal"
+              min="10.0"
+              value="25.0"
+              aria-valuetext="25.0°C"
+              max="38.0"
+              step="0.1"
+              aria-orientation="vertical">
+          </div>
+        </div>
+      </div>
+      <div role="separator" id="ex_end_sep" aria-labelledby="ex_end_sep ex_label" aria-label="End of"></div>
+    </section>
+
+      <section>
+        <h2>Accessibility Features</h2>
+        <ul>
+          <li>
+            A border around the thumb is used for indicating the range control with keyboard focus.
+          </li>
+        </ul>
+      </section>
+
+  <section>
+      <h2 id="rps_label">Role, Property, State, and Tabindex Attributes</h2>
+
+      <table aria-labelledby="rps_label" class="data attributes">
+        <thead>
+          <tr>
+            <th scope="col">Role</th>
+            <th scope="col">Attribute</th>
+            <th scope="col">Element</th>
+            <th scope="col">Usage</th>
+          </tr>
+        </thead>
+        <tbody>
+          <tr data-test-id="range-aria-valuetext">
+            <td></td>
+            <th scope="row">
+              <code>aria-valuetext</code>
+            </th>
+            <td>
+              <code>input[type=<q>range</q>]</code>
+            </td>
+            <td>
+              <ul>
+                <li>A string value that provides a user-friendly name for the current value of the range.</li>
+                <li>For the temperature range, is the value of the range appended with the text <q>°C</q>.</li>
+                <li>For fan and heat/cool, is the text representation of the current state, e.g., <q>Off</q> or <q>Heat</q>.</li>
+              </ul>
+            </td>
+          </tr>
+          <tr data-test-id="range-aria-orientation">
+            <td></td>
+            <th scope="row">
+              <code>aria-orientation=<q>vertical</q></code>
+            </th>
+            <td>
+              <code>input[type=<q>range</q>]</code>
+            </td>
+            <td>
+              <ul>
+                <li>Indicates the orientation of the <code>input</code> element.</li>
+                <li>Set to <code>vertical</code> for the temperature range control.</li>
+              </ul>
+            </td>
+          </tr>
+        </tbody>
+      </table>
+
+    <section>
+      <h2>Javascript and CSS Source Code</h2>
+      <ul id="css_js_files">
+        <li>CSS: <a href="css/range-temperature.css" type="text/css">range-temperature.css</a></li>
+        <li>Javascript: <a href="js/range-temperature.js" type="text/javascript">range-temperature.js</a></li>
+      </ul>
+    </section>
+
+    <section>
+      <h2 id="sc1_label">HTML Source Code</h2>
+      <div role="separator" id="sc1_start_sep" aria-labelledby="sc1_start_sep sc1_label" aria-label="Start of"></div>
+      <pre><code id="sc1"></code></pre>
+      <div role="separator" id="sc1_end_sep" aria-labelledby="sc1_end_sep sc1_label" aria-label="End of"></div>
+      <script>
+        sourceCode.add('sc1', 'ex1', 'ex_label', 'css_js_files');
+        sourceCode.make();
+      </script>
+    </section>
+  </main>
+  <nav>
+    <a href="../../#slider">Slider Design Pattern in WAI-ARIA Authoring Practices 1.2</a>
+  </nav>
+</body>
+</html>
